refactor: isolate credit notification config actions
